Rename utils.js
